Refactor App component to remove ProfilePage and related logic; update DashboardPage rendering. Simplify EnhancedAgentCockpit import path. Enhance dashboard components index with documentation and export consistency. Update MCP hooks to use ES6 imports and exports; alias useMcpGameState as useGameState. Remove placeholder widget components from DashboardPage for cleaner code.
