+++ apps/web/src/pages/Dashboard.jsx
@@ -131,97 +131,6 @@ const DashboardPage = () => {
     localStorage.setItem('dashboard-layout', JSON.stringify(newLayout));
   };
   
-  // Create placeholder components for missing widgets
-  const WorkspaceWidget = () => (
-    <div className="placeholder-widget">
-      <h3>KI-Workspace</h3>
-      <p>Workspace-Komponente wird geladen...</p>
-      <style jsx>{`
-        .placeholder-widget {
-          padding: 2rem;
-          text-align: center;
-          background-color: var(--surface-color);
-          border-radius: 8px;
-          box-shadow: 0 2px 10px var(--shadow-color);
-          height: 100%;
-          display: flex;
-          flex-direction: column;
-          justify-content: center;
-          align-items: center;
-        }
-        
-        h3 {
-          margin-bottom: 1rem;
-          color: var(--primary-color);
-        }
-        
-        p {
-          color: var(--text-secondary-color);
-        }
-      `}</style>
-    </div>
-  );
-  
-  const SupportWidget = () => (
-    <div className="placeholder-widget">
-      <h3>Starthilfe & Support</h3>
-      <p>Support-Komponente wird geladen...</p>
-      <style jsx>{`
-        .placeholder-widget {
-          padding: 2rem;
-          text-align: center;
-          background-color: var(--surface-color);
-          border-radius: 8px;
-          box-shadow: 0 2px 10px var(--shadow-color);
-          height: 100%;
-          display: flex;
-          flex-direction: column;
-          justify-content: center;
-          align-items: center;
-        }
-        
-        h3 {
-          margin-bottom: 1rem;
-          color: var(--primary-color);
-        }
-        
-        p {
-          color: var(--text-secondary-color);
-        }
-      `}</style>
-    </div>
-  );
-  
-  const NewsFeedWidget = () => (
-    <div className="placeholder-widget">
-      <h3>Neues aus dem AGENT_LAND</h3>
-      <p>News-Feed-Komponente wird geladen...</p>
-      <style jsx>{`
-        .placeholder-widget {
-          padding: 2rem;
-          text-align: center;
-          background-color: var(--surface-color);
-          border-radius: 8px;
-          box-shadow: 0 2px 10px var(--shadow-color);
-          height: 100%;
-          display: flex;
-          flex-direction: column;
-          justify-content: center;
-          align-items: center;
-        }
-        
-        h3 {
-          margin-bottom: 1rem;
-          color: var(--primary-color);
-        }
-        
-        p {
-          color: var(--text-secondary-color);
-        }
-      `}</style>
-    </div>
-  );
-  
   return (
     <div className="dashboard-page">
       <Dashboard
